separate into core mgnifier and resource proxies
